rename: `strongbox` to `strong_box`
